+++ src/Unitful.jl
@@ -20,8 +20,8 @@ import LinearAlgebra: Diagonal, Bidiagonal, Tridiagonal, SymTridiagonal
 import LinearAlgebra: istril, istriu, norm
 import Random
 
-export logunit, unit, dimension, uconvert, ustrip, upreferred
-export @dimension, @derived_dimension, @refunit, @unit, @u_str
+export logunit, unit, absoluteunit, dimension, uconvert, ustrip, upreferred
+export @dimension, @derived_dimension, @refunit, @unit, @affineunit, @u_str
 export Quantity, DimensionlessQuantity, NoUnits, NoDims
 
 export uconvertp, uconvertrp, convertr, convertrp, reflevel, linear