+++ src/quantities.jl
@@ -27,7 +27,11 @@ Quantity(x::Number, y::Units{()}) = x
 *(x::Quantity, y::Bool) = Quantity(x.val*y, unit(x))
 
 *(y::Number, x::Quantity) = *(x,y)
-*(x::Quantity, y::Number) = Quantity(x.val*y, unit(x))
+function *(x::Quantity, y::Number)
+    x isa AffineQuantity &&
+        throw(AffineError("an invalid operation was attempted with affine quantities: $x*$y"))
+    return Quantity(x.val*y, unit(x))
+end
 
 *(A::Units, B::AbstractArray) = broadcast(*, A, B)
 *(A::AbstractArray, B::Units) = broadcast(*, A, B)
@@ -77,7 +81,6 @@ for op in [:+, :-]
     @eval ($op)(x::Quantity{S,D,U}, y::Quantity{T,D,U}) where {S,T,D,U} =
         Quantity(($op)(x.val, y.val), U())
 
-    # If not generated, there are run-time allocations
     @eval function ($op)(x::Quantity{S,D,SU}, y::Quantity{T,D,TU}) where {S,T,D,SU,TU}
         ($op)(promote(x,y)...)
     end
@@ -85,6 +88,36 @@ for op in [:+, :-]
     @eval ($op)(x::Quantity, y::Quantity) = throw(DimensionError(x,y))
     @eval ($op)(x::Quantity) = Quantity(($op)(x.val), unit(x))
 end
+
+function +(x::AffineQuantity{S,D}, y::Quantity{T,D}) where {S,T,D}
+    pu = promote_unit(unit(x), unit(y))     # units for the final result.
+
+    # Get x on an absolute scale. FreeUnits in the line below prevents
+    # promote(x′, y) from yielding affine quantities. If x had `ContextUnits` and
+    # the promotion units were affine units, x′+y would error without this.
+    x′ = Quantity(x.val - affinetranslation(unit(x)), FreeUnits(absoluteunit(x)))
+
+    # Likewise if y were not affine but y had ContextUnits and the promotion units were
+    # affine, x′+y could also fail.
+    y′ = Quantity(y.val, FreeUnits(unit(y)))
+
+    return uconvert(pu, x′+y′)  # we get back the promotion context in the end
+end
++(x::Quantity, y::AffineQuantity) = +(y,x)
+
+# Disallow addition of affine quantities
++(x::AffineQuantity, y::AffineQuantity) = throw(AffineError(
+   "an invalid operation was attempted with affine quantities: $x + $y"))
+
+# Specialize substraction of affine quantities
+-(x::AffineQuantity, y::AffineQuantity) = -(promote(x,y)...)
+function -(x::T, y::T) where T <: AffineQuantity
+    return Quantity(x.val - y.val, absoluteunit(unit(x)))
+end
+
+# Disallow subtracting an affine quantity from a quantity
+-(x::Quantity, y::AffineQuantity) =
+    throw(AffineError("an invalid operation was attempted with affine quantities: $x - $y"))
 
 # Needed until LU factorization is made to work with unitful numbers
 function inv(x::StridedMatrix{T}) where {T <: Quantity}
@@ -253,13 +286,19 @@ for f in (:floor, :ceil, :trunc, :round)
 end
 
 zero(x::Quantity) = Quantity(zero(x.val), unit(x))
-zero(x::Type{Quantity{T,D,U}}) where {T,D,U} = zero(T)*U()
+zero(x::AffineQuantity) = Quantity(zero(x.val), absoluteunit(x))
+zero(x::Type{Quantity{T,D,U}}) where {T,D,U<:ScalarUnits} = zero(T)*U()
+zero(x::Type{Quantity{T,D,U}}) where {T,D,U<:AffineUnits} = zero(T)*absoluteunit(U())
 
 one(x::Quantity) = one(x.val)
+one(x::AffineQuantity) =
+    throw(AffineError("no multiplicative identity for affine quantity $x."))
 get_T(::Type{Quantity{T}}) where T = T
 get_T(::Type{Quantity{T,D}}) where {T,D} = T
 get_T(::Type{Quantity{T,D,U}}) where {T,D,U} = T
 one(x::Type{<:Quantity}) = one(get_T(x))
+one(x::Type{<:AffineQuantity}) =
+    throw(AffineError("no multiplicative identity for affine quantity type $x."))
 
 isreal(x::Quantity) = isreal(x.val)
 isfinite(x::Quantity) = isfinite(x.val)